tests/frank-marco/needsreview/contacts/contacts.html

<!DOCTYPE html>
<html>
<head>
    <title>Contacts API Specs</title>
    <meta name="viewport" content="width=device-width, height=device-height, user-scalable=yes, initial-scale=1.0;"/>

    <!-- Test Runner -->
    <script src="cordova-2.1.0.js"></script>
    <script src="resources/testharness.js"></script>
    <script src="resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<!--
REMOVE THIS

TESTHARNESS.JS API:

  http://hoppipolla.co.uk/talks/ttwf/testharness.html

IMPORTANT:

  Create a symbolic link from your hard drive root directory to the
  Dropbox /HTML5/ so that the testharness.js reference above works
  without changes.
  For example:
    sudo ln -s ~/Dropbox/TestTWF/MOBILE/resources /resources
-->
<script type="text/javascript">

setup({explicit_done:true});

document.addEventListener("deviceready", function () {

    /* Here put you (a)synchronous tests */

    var gContactObj = null;
    var removeContact = function () {
        if (gContactObj) {
            gContactObj.remove(function () {
            }, function () {
                console.log("[CONTACTS ERROR]: removeContact cleanup method failed to clean up test artifacts.");
            });
            gContactObj = null;
        }
    };
    /* Contacts (navigator.contacts) match WebIDL */

    test(function () {
        assert_idl_attribute(navigator, 'contacts', '');
    }, "Contacts (navigator.contacts) match WebIDL");

    /* Contacts (navigator.contacts) should exist */

    test(function () {
        assert_true(typeof navigator.contacts !== 'undefined');
        assert_true(typeof navigator.contacts === 'object');

        /* should contain a find function */

        test(function () {
            assert_true(typeof navigator.contacts.find !== 'undefined');
            assert_true(typeof navigator.contacts.find === 'function');
            assert_own_property(navigator.contacts, 'find', '');
        }, "should contain a find function");

        /* has Object ContactFindOptions */

        test(function () {
            assert_true(ContactFindOptions !== 'undefined');
            assert_true(typeof ContactFindOptions === 'function');
            assert_own_property(window, 'ContactFindOptions', '');
        }, "has Object ContactFindOptions");

        /*contacts should find something and call success callback*/

        var findTest = async_test("success callback should be called with an array");
        var fail = findTest.step_func(function () {
            assert_unreached("find failed");
        });
        var win = findTest.step_func(function (result) {
            assert_true(result !== 'undefined');
            assert_true(Array.isArray(result));
            findTest.done();
        });

        findTest.step(function () {
            obj = new ContactFindOptions();
            obj.filter = "";
            obj.multiple = true;
            navigator.contacts.find(["displayName", "name", "phoneNumbers", "emails"], win, fail, obj);
        });

        /*should throw an exception if success callback is empty*/

        // what kind of exception?!

        test(function () {
            assert_throws("NOT_FOUND_ERR", function () {
                var fail = function () {
                };
                var obj = new ContactFindOptions();
                obj.filter = "";
                obj.multiple = true;
                navigator.contacts.find(["displayName", "name", "phoneNumbers", "emails"], null, fail, obj);
            });
        }, "should throw an exception if success callback is empty");

        /* error callback should be called when no fields are specified */

        var findTest2 = async_test("error callback should be called when no fields are specified");
        var fail = findTest2.step_func(function () {
            findTest2.done();
        });
        var win = findTest2.step_func(function (result) {
            assert_unreached("success callback was called");
        });

        findTest2.step(function () {
            obj = new ContactFindOptions();
            obj.filter = "";
            obj.multiple = true;
            navigator.contacts.find([], win, fail, obj);
        });

    }, "Contacts (navigator.contacts) should exist");

    /* with newly-created contact */

    test(function () {
        removeContact();

        var createPart = async_test("create a new contact");

        var test = createPart.step_func(function () {
            var find = async_test("should be able to find a contact by name");

            var foundName = function (result) {
                var bFound = false;
                try {
                    for (var i = 0; i < result.length; i++) {
                        if (result[i].name.familyName == "Delete") {
                            bFound = true;
                            break;
                        }
                    }
                } catch (e) {
                    return false;
                }
                return bFound;
            };

            var findWin = find.step_func(function (result) {
                assert_true(result !== 'undefined', 'result is undefined');
                assert_true(Array.isArray(result), 'result is not an array');
                assert_true(result.length > 1, 'result length is smaller than 1');
                assert_true(foundName(result), 'foundName has not found any contact');
                find.done();
            });

            var findFail = find.step_func(function (result) {
                assert_unreached("find the new created contact failed");
            });

            find.step(function () {
                obj = new ContactFindOptions();
                obj.filter = "Delete";
                obj.multiple = true;
                navigator.contacts.find(["displayName", "name", "phoneNumbers", "emails"], findWin, findFail, obj);
            });
            createPart.done();
        });

        var fail = createPart.step_func(function (result) {
            assert_unreached("creating a contact failed");
        });

        createPart.step(function () {
            gContactObj = new Contact();
            gContactObj.name = new ContactName();
            gContactObj.name.familyName = "Delete";
            gContactObj.save(test, fail);
        });


    }, "find newly-created contact");

    var contact;
    test(function () {
        contact = new Contact("a", "b", new ContactName("a", "b", "c", "d", "e", "f"), "c", [], [], [], [], [], "f", "i", [], [], []);
        assert_true(typeof contact !== "undefined");

        test_Contact(contact);

    }, "should be able to create a Contact instance");

    function test_Contact(contact) {
        test(function () {
            assert_true(typeof contact.save !== "undefined");
            assert_true(typeof contact.save === "function");
        }, "Contact.save should be a function");

        test(function () {
            assert_true(typeof contact.remove !== "undefined");
            assert_true(typeof contact.remove === "function");
        }, "Contact.remove should be a function");

        test(function () {
            assert_true(contact.id === "a");
        }, "Contact.id should be 'a'");

        test(function () {
            assert_true(contact.displayName === "b");
        }, "Contact.displayName should be 'b'");

        test(function () {
            assert_true(contact.name.formatted === "a");
        }, "Contact.name.formatted should be 'a'");

        test(function () {
            assert_true(contact.nickname === "c");
        }, "Contact.nickname should be 'c'");

        test(function () {
            assert_true(typeof contact.phoneNumbers !== "undefined");
        }, "Contact.phoneNumbers should be exist");

        test(function () {
            assert_true(typeof contact.emails !== "undefined");
        }, "Contact.emails should be exist");

        test(function () {
            assert_true(typeof contact.addresses !== "undefined");
        }, "Contact.addresses should be exist");

        test(function () {
            assert_true(typeof contact.ims !== "undefined");
        }, "Contact.ims should be exist");

        test(function () {
            assert_true(typeof contact.organizations !== "undefined");
        }, "Contact.organizations should be exist");

        test(function () {
            assert_true(contact.birthday === "f");
        }, "Contact.birthday should be 'f'");

        test(function () {
            assert_true(contact.note === "i");
        }, "Contact.birthday should be 'i'");

        test(function () {
            assert_true(typeof contact.photos !== "undefined");
        }, "Contact.photos should be exist");

        test(function () {
            assert_true(typeof contact.categories !== "undefined");
        }, "Contact.categories should be exist");

        test(function () {
            assert_true(typeof contact.urls !== "undefined");
        }, "Contact.urls should be exist");
    }

    var contactName;
    test(function () {
        contactName = new ContactName("Dr. First Last Jr.", "Last", "First", "Middle", "Dr.", "Jr.");
        assert_true(typeof contactName !== "undefined");

        test_ContactName(contactName);

    }, "should be able to create a ContactName instance");

    function test_ContactName(contactName) {
        test(function () {
            assert_true(contactName.formatted === "Dr. First Last Jr.");
        }, "ContactName.formatted should be 'Dr. First Last Jr.'");

        test(function () {
            assert_true(contactName.familyName === "Last");
        }, "ContactName.familyName should be 'Last'");

        test(function () {
            assert_true(contactName.givenName === "First");
        }, "ContactName.givenName should be 'First'");

        test(function () {
            assert_true(contactName.middleName === "Middle");
        }, "ContactName.middleName should be 'Middle'");

        test(function () {
            assert_true(contactName.honorificPrefix === "Dr.");
        }, "ContactName.honorificPrefix should be 'Dr.'");

        test(function () {
            assert_true(contactName.honorificSuffix === "Jr.");
        }, "ContactName.honorificSuffix should be 'Jr.'");
    }

    var contactField;
    test(function () {
        contactField = new ContactField("home", "8005551212", true);
        assert_true(typeof contactField !== "undefined");

        test_ContactField(contactField);

        var field = new ContactField(12345678, 12345678, true);

        test(function () {
            assert_true(field.type === "12345678");
        }, "ContactField.type should be converted to the string '12345678'");

        test(function () {
            assert_true(field.value === "12345678");
        }, "ContactField.value should be converted to the string '12345678'");

    }, "should be able to create a ContactField instance");

    function test_ContactField(contactField) {
        test(function () {
            assert_true(contactField.type === "home");
        }, "ContactField.type should be 'home'");

        test(function () {
            assert_true(contactField.value === "8005551212");
        }, "ContactField.value should be '8005551212'");

        test(function () {
            assert_true(contactField.pref === true);
        }, "ContactField.pref should be true");
    }

    var contactAddress;
    test(function () {
        contactAddress = new ContactAddress(true, "home", "a", "b", "c", "d", "e", "f");
        assert_true(typeof contactAddress !== "undefined");

        test_ContactAddress(contactAddress);

    }, "should be able to create a ContactAddress instance");

    function test_ContactAddress(contactAddress) {
        test(function () {
            assert_true(contactAddress.pref === true);
        }, "ContactAddress.pref should be true");

        test(function () {
            assert_true(contactAddress.type === "home");
        }, "ContactAddress.type should be 'home'");

        test(function () {
            assert_true(contactAddress.formatted === "a");
        }, "ContactAddress.formatted should be 'a'");

        test(function () {
            assert_true(contactAddress.streetAddress === "b");
        }, "ContactAddress.streetAddress should be 'b'");

        test(function () {
            assert_true(contactAddress.locality === "c");
        }, "ContactAddress.locality should be 'c'");

        test(function () {
            assert_true(contactAddress.region === "d");
        }, "ContactAddress.region should be 'd'");

        test(function () {
            assert_true(contactAddress.postalCode === "e");
        }, "ContactAddress.postalCode should be 'e'");

        test(function () {
            assert_true(contactAddress.country === "f");
        }, "ContactAddress.country should be 'f'");
    }

    var contactOrg;
    test(function () {
        contactOrg = new ContactOrganization(true, "home", "a", "b", "c", "d", "e", "f", "g");
        assert_true(typeof contactOrg !== "undefined");

        test_ContactOrganization(contactOrg);

    }, "should be able to create a ContactOrganization instance");

    function test_ContactOrganization(contactOrg) {

        test(function () {
            assert_true(contactOrg.pref === true);
        }, "ContactOrg.pref should be true");

        test(function () {
            assert_true(contactOrg.type === "home");
        }, "ContactOrg.type should be 'home'");

        test(function () {
            assert_true(contactOrg.name === "a");
        }, "ContactOrg.name should be 'a'");

        test(function () {
            assert_true(contactOrg.department === "b");
        }, "ContactOrg.department should be 'b'");

        test(function () {
            assert_true(contactOrg.title === "c");
        }, "ContactOrg.title should be 'c'");
    }

    test(function () {
        var contactFindOptions = new ContactFindOptions("a", true, "b");
        assert_true(typeof contactFindOptions !== "undefined");

        test(function () {
            assert_true(contactFindOptions.filter === 'a');
        }, "ContactFindOptions.filter should be 'a'");

        test(function () {
            assert_true(contactFindOptions.multiple === true);
        }, "ContactFindOptions.multiple should be true");

    }, "should be able to create a ContactFindOptions instance");

    test(function () {
        var contact = new Contact();
        assert_true(typeof contact.clone === "function");

    }, "Contact.clone should be function");

    test(function () {
        var contact = new Contact();
        contact.id = 1;
        contact.displayName = "Test Name";
        contact.nickname = "Testy";
        contact.gender = "male";
        contact.note = "note to be cloned";
        contact.name = new ContactName("Mr. Test Name");

        var clonedContact = contact.clone();
        assert_true(typeof clonedContact !== "undefined");

        test(function () {
            assert_true(contact.id === 1);
        }, "Contact.id should be 1");

        test(function () {
            assert_true(clonedContact.id === null);
        }, "ClonedContact.id should be null");

        test(function () {
            assert_true(clonedContact.displayName === contact.displayName);
        }, "ClonedContact.displayName should be contact.displayName");

        test(function () {
            assert_true(clonedContact.nickname === contact.nickname);
        }, "ClonedContact.nickname should be contact.nickname");

        test(function () {
            assert_true(clonedContact.gender === contact.gender);
        }, "ClonedContact.gender should be contact.gender");

        test(function () {
            assert_true(clonedContact.note === contact.note);
        }, "ClonedContact.note should be contact.note");

        test(function () {
            assert_true(clonedContact.name.formatted === contact.name.formatted);
        }, "ClonedContact.name.formatted should be contact.name.formatted");

        test(function () {
            assert_true(clonedContact.connected === contact.connected);
        }, "ClonedContact.connected should be contact.connected");

    }, "Contact.clone function should make deep copy of Contact Object");

    test(function () {
        assert_true(typeof navigator.contacts.create !== "undefined");
        assert_true(typeof navigator.contacts.create === "function");
    }, "Contacts (navigator.contacts.create) should a function");

    test(function () {
        var bDay = new Date(1976, 7, 4);
        var contact = navigator.contacts.create({"displayName":"test name", "gender":"male", "note":"my note", "name":{"formatted":"Mr. Test Name"}, "emails":[
            {"value":"here@there.com"},
            {"value":"there@here.com"}
        ], "birthday":bDay});
        assert_true(typeof contact !== "undefined");

        test(function () {
            assert_true(contact.displayName === 'test name');
        }, "Contact.displayName should hold 'test name'");

        test(function () {
            assert_true(contact.note === 'my note');
        }, "Contact.note should hold 'my note'");

        test(function () {
            assert_true(contact.name.formatted === 'Mr. Test Name');
        }, "Contact.name.formatted should hold 'Mr. Test Name'");

        test(function () {
            assert_true(contact.emails.length === 2);
        }, "Contact.emails should hold 2 emails");

        test(function () {
            assert_true(contact.emails[0].value === 'here@there.com');
        }, "Contact.emails[0] should hold 'here@there.com'");

        test(function () {
            assert_true(contact.emails[1].value === 'there@here.com');
        }, "Contact.emails[1] should hold 'there@here.com'");

        test(function () {
            assert_true(contact.nickname === null);
        }, "Contact.nickname should be null");

        test(function () {
            var bDay = new Date(1976, 7, 4);
            assert_true(contact.birthday.getTime() == bDay.getTime());
        }, "Contact.birthday should be bDay");

    }, "Contacts.create should return a Contact object");
    /* End your tests here */

    // Now it's time to call the final done()
    setTimeout(function(){
        done();
    }, 1000);
}, false);


</script>
</body>
</html>
